Fix error with older version of iOS
Fix issue that can occur with an older version of iOS.

+++ Sources/SwiftletRadioButtonPicker/SwiftletRadioButtonPicker.swift
@@ -150,7 +150,7 @@ public struct SwiftletRadioButtonPicker: View {
         HStack {
             Image(systemName: (button.isSelected) ? selectedSymbolName : unselectedSymbolName)
                 .resizable()
-                .frame(width: selectorSize, height: selectorSize)
+                .frame(width: CGFloat(selectorSize), height: CGFloat(selectorSize))
             
             Text(button.name)
